+++ templates/default/src/include/src/components/FeedItems.vue
@@ -1,5 +1,5 @@
 <script setup>
-import { ref, watchEffect, shallowRef, onMounted, watch } from 'vue'
+import { ref, shallowRef, onMounted, watch } from 'vue'
 import { storeToRefs } from 'pinia'
 import { useEmbedStore } from '../stores/embed'
 import { useAudioStore } from '../stores/audio'
@@ -55,12 +55,16 @@ const props = defineProps({
 
 const filteredFeedItems = shallowRef([])
 const initialized = ref(false)
-const emit = defineEmits(['expand-article', 'unexpand-article', 'expand-feed', 'unexpand-feed'])
+const emit = defineEmits([
+  'expand-article',
+  'unexpand-article',
+  'expand-feed',
+  'unexpand-feed',
+  'feed-loading',
+  'feed-loaded',
+])
 const itemDetails = ref(null)
 
-fStore.$subscribe((state) => {
-  filterFeedItems(state.payload)
-})
 
 // Utility
 // ===============
@@ -98,6 +102,7 @@ const aggregateItems = (items) => {
     }
     result[d].push(item)
   }
+  emit('feed-loaded', items.length)
   return result
 }
 
@@ -182,6 +187,7 @@ const retrieveItemData = async () => {
   return await getFeedItems(props.feed.id, props.archived)
 }
 const reload = async () => {
+  emit('feed-loading')
   if (!initialized.value) {
     await retrieveItemData()
   }
@@ -191,11 +197,10 @@ const reload = async () => {
   } else {
     filteredFeedItems.value = aggregateItems(await retrieveItemData())
   }
-
 }
 // ===================
-watchEffect(async () => {
-  reload()
+fStore.$subscribe(async () => {
+  await reload()
 })
 
 watch(feedReloadTrigger, () => {
@@ -240,7 +245,8 @@ const updateArticleHighlighting = () => {
   }
 }
 
-onMounted(() => {
+onMounted(async () => {
+  await reload()
   setInterval(() => {
     updateArticleHighlighting()
   }, 400)